place-recommender-agent/pyproject.toml
[project]
name = "place-recommender-agent"
version = "0.1.0"
description = "Place recommender agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic-ai>=1.88.0,<2.0.0",
    "pydantic-settings>=2.14.0,<3.0.0",
    "fasta2a>=0.6.0,<1.0.0",
    "logfire>=3.0.0",
]
